+++ packages/client/src/features/ConnectionView.tsx
@@ -1,0 +1,9 @@
+import { NetworkAdvancedPanel } from "./NetworkAdvancedPanel.js";
+
+export function ConnectionView() {
+  return (
+    <div className="stack">
+      <NetworkAdvancedPanel />
+    </div>
+  );
+}