+++ packages/client/src/vite-env.d.ts
@@ -1,6 +1,7 @@
 /// <reference types="vite/client" />
 /// <reference types="vite-plugin-pwa/client" />
 
+declare const __APP_VERSION__: string;
 declare const __PROOF_E2E__: "0" | "1";
 
 interface ImportMetaEnv {